requirements.txt
azure-devops>=7.1.0b4
python-dotenv>=1.0.0
google-generativeai>=0.3.0
requests>=2.31.0
flask>=3.0.0
python-dateutil>=2.9.0
